[CLI] Move dep install to a shared directory
