+++ packages/cli/src/generate/utils/index.ts
@@ -3,3 +3,4 @@ export { rmDirAndFilesIfExist } from './rm-dir-and-files-if-exist';
 export { mkdirIfDoesNotExist } from './mkdir-if-does-not-exist';
 export { getNames } from './get-names';
 export { logger } from './logger';
+export { installDependencies } from './install-dependencies';
